Add systemd system service
Allows installation of a canonical systemd service that starts and supervises
waydroid-sensord.

This file is taken from our downstream services package in postmarketOS
and upstreaming it could make future improvements easier and downstream
packaging.

I'm not the best at CMake configuration, so please let me know if there
is something to nag about.